refactor: Restructure app initialization and enhance environment variable handling. Closes #30
