+++ templates/express-mongo/.env.example
@@ -1,0 +1,1 @@
+MONGO_URI=
