+++ templates/express-mongo/package.json
@@ -3,14 +3,13 @@
   "version": "1.0.0",
   "type": "module",
   "scripts": {
-    "build": "tsc && node fix-imports.js",
-    "start": "node build/app.js",
-    "dev": "ts-node src/server.ts"
+    "dev": "tsx watch src/server.ts"
   },
   "dependencies": {
     "dotenv": "^16.3.1",
     "express": "^4.18.2",
-    "mongoose": "^8.0.3"
+    "mongoose": "^8.0.3",
+    "cors": "^2.8.5"
   },
   "devDependencies": {
     "@types/express": "^4.17.21",
