+++ templates/express-mongo/src/config/database.ts
@@ -1,7 +1,7 @@
 import mongoose from "mongoose";
-import { config } from "dotenv";
+import dotenv from "dotenv";
 
-config();
+dotenv.config();
 
 const MONGO_URI = process.env.MONGO_URI as string;
